+++ Mammal.java
@@ -12,7 +12,7 @@ public class Mammal extends Animal
 	public int getNumberOfArms() { return NumberOfArms; }
 	public int getNumberOfLegs() { return NumberOfLegs; }
 
-	public Mammal(int Age, String Consevervation_status, String Diet, boolean Flight, String Habit, boolean healthy, String Last_Checkup, int Lifespan, String Location, String Name, int Weight, int NumberOfArms, int NumberOfLegs){
+	public Mammal(int Age, String Consevervation_status, String Diet, boolean Flight, String Habit, boolean healthy, String Last_Checkup, int Lifespan, String Location, String Name, double Weight, int NumberOfArms, int NumberOfLegs){
 		super(Age, Consevervation_status, Diet, Flight,  Habit, healthy, Last_Checkup, Lifespan, Location, Name, Weight);
 		this.NumberOfArms = NumberOfArms;
 		this.NumberOfLegs = NumberOfLegs;
